Create basic tic-tac-toe game for two users
The angular page now allows two users to play a game of tic-tac-toe against each other on the same device. Clicking a cell sets it to be either an 'X' or an 'O', depending on whose turn it is. The turns alternate after each move is made. Once a cell has been clicked (marked with an X or O), it's value cannot change. The page checks for a win in every direction for either player and does not allow making marks after the game has been won.

To do:
-Draw a line showing where the game was won
-Add the option to switch to play against a bot
-Implement the bot