Add settings menu with reset button and bot settings
There is now a settings menu that has a button to reset the game. Also in the settings is a switch to allow the user to choose between playing the bot and playing against themselves (another user on the same device). When the play bot switch is set to true, there is another switch that allows the user to choose to have the bot go first or second.